requirements.txt
# requirements.txt
torch==2.0.1
torchvision==0.15.2
mmsegmentation==1.0.0
opencv-python==4.7.0.72
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
albumentations==1.3.1
tqdm==4.65.0
scikit-learn==1.2.2
pandas==2.0.2
tensorboard
pillow
